Updated to reflect Carolyn Brown's request to show type against event.
“Hi Guy

I’m looking at the changes we need to show events other than tutorials
(but not residential schools) on Student Home / Tutor Home.

We thought it would be helpful to specify the format of the event
(tutorial, forum, field trip) on the Tutorials for module page on
Student Home as we cannot rely on the event title being
self-explanatory. Would you be able to amend the design?”